Move interfaces back into charm.py for now to clean up PR
